64-bit content-length
git-svn-id: svn://svn.caucho.com/home/svn/svnroot/resin/trunk@9785 9c94448d-38f1-0310-a231-d98308ff1ebf

+++ modules/resin/src/com/caucho/server/http/AbstractHttpResponse.java
@@ -708,7 +708,7 @@ abstract public class AbstractHttpResponse {
    *
    * @param length the length of the content.
    */
-  public void setContentLength(int length)
+  public void setContentLength(long length)
   {
     _contentLength = length;
   }
